do not prevent triggering change event on setting initial prop value
This allows a prop bound to a computed property on the child vm to get its
initial value from the (child's) setter function -- e.g. casting the value
received from parent, or setting a default value.

Caveat: the parent vm should support setting this initial value without it
causing this child vm to be destroyed and recreated.